requirements.txt
google-genai==0.3.0
yfinance>=1.2.0
markdown==3.5.2
python-dateutil==2.8.2
urllib3<2
